refactor(deps): remove arrow lib dependencies
- split logout e2e test
