+++ buildSrc/src/main/kotlin/kotlin-conventions.gradle.kts
@@ -75,7 +75,6 @@ dependencies {
     implementation(kotlin("reflect"))
     add("implementation", libs.findBundle("kotlinx-coroutines").get())
     add("implementation", libs.findLibrary("kotlin-reflect").get())
-    add("implementation", libs.findBundle("arrow").get())
     // Align versions of all Kotlin components
     implementation(platform("org.jetbrains.kotlin:kotlin-bom"))
